初步完成
1.重写了单词选择
2.整理文件
