--- Palabra/dict_txt_simple.txt
@@ -1,8 +1,0 @@
-dad
-n
-°Ö°Ö
-I have a pretty dad.
-son
-n
-¶ù×Ó
-All of my sons live with me.
