+++ Palabra/res/data_zly.json
@@ -1,0 +1,1 @@
+{"dict":[{"Word":"father","err":0,"fav":false,"imp":0,"last":1496405701,"note":"It's my favortie word!","rev":0}],"history":[{"cou":999,"day":1,"mon":6,"year":2016},{"cou":999,"day":1,"mon":6,"year":2016},{"cou":999,"day":1,"mon":6,"year":2016}],"lev":"CET6","num":0}
