--- Palabra/user_simple.ini
@@ -1,16 +1,0 @@
-zhengly
-aabbcc
-yangle
-yangle
-yl
-yl
-yyl
-yyl
-yangyangle
-yangyangle
-yyle
-yyle
-lala
-lala
-hkz
-hkz
