+++ Palabra/user_word.cpp
@@ -22,3 +22,4 @@ std::ostream &operator<<(std::ostream & os, const UserWord & user_word)
 	//}
 	return os;
 }
+
