--- Palabra/userdata_zly.txt
@@ -1,4 +1,0 @@
-father
-This is the most important word!
-Favorite
-1 1 111 5
