--- Palabra/userlist.json
@@ -1,6 +1,0 @@
-[
-	{
-		"Password" : "111111",
-		"Username" : "zhengly123"
-	}
-]
