--- Palabra/word(empty).h
@@ -1,10 +1,0 @@
-/*#pragma once
-#ifndef WORD_H
-#define WORD_H
-#include<string>
-struct Word
-{
-	
-};
-
-#endif  // !WORD_H*/
